improve playlist ui

+++ Midibard/Playlist/Repositories/ISongRepository.cs
@@ -16,7 +16,7 @@ public interface ISongRepository
 
     // Song-specific operations
     Task IncrementPlayCountAsync(int songId);
-    Task SetRatingAsync(int songId, double rate);
+    Task SetRatingAsync(int songId, int rate);
     Task AddTagAsync(int songId, string tag);
     Task RemoveTagAsync(int songId, string tag);
 }
